+++ src/app/authorize/pages/authorize/authorize.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core'
+import { take } from 'rxjs/operators'
+import { PlatformInfo, PlatformInfoService } from 'src/app/cdk/platform-info'
 import { UserService } from 'src/app/core'
+import { TogglzService } from 'src/app/core/togglz/togglz.service'
 
 @Component({
   templateUrl: './authorize.component.html',
@@ -7,8 +10,15 @@ import { UserService } from 'src/app/core'
   preserveWhitespaces: true,
 })
 export class AuthorizeComponent implements OnInit {
+  platform: PlatformInfo
   showAuthorizationComponent: boolean
-  constructor(_user: UserService) {
+  signInUpdatesV1Togglz = false
+
+  constructor(
+    _user: UserService,
+    private _platformInfo: PlatformInfoService,
+    private _togglz: TogglzService
+  ) {
     _user.getUserSession().subscribe((session) => {
       if (session.oauthSession && session.oauthSession.error) {
         this.showAuthorizationComponent = false
@@ -16,7 +26,15 @@ export class AuthorizeComponent implements OnInit {
         this.showAuthorizationComponent = true
       }
     })
+    _platformInfo.get().subscribe((platformInfo) => {
+      this.platform = platformInfo
+    })
   }
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this._togglz
+      .getStateOf('SIGN_IN_UPDATES_V1')
+      .pipe(take(1))
+      .subscribe((value) => (this.signInUpdatesV1Togglz = value))
+  }
 }